# Remove custom SortedList<T> enumerator