The `List<T>` enumerator has additional safety guards preventing against modifying the collection during enumeration. There was no reason for `SortedList<T>` to have a custom enumerator other than to prevent boxing, but `List<T>.Enumerator` is already exposed for that purpose.